% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbturns.R
\name{bbturns}
\alias{bbturns}
\title{turns of a time series}
\usage{
bbturns(x, neighbour = 5, mincycle = 15, phase = 5, extr_limit = 3.5,
  end_num = 6, mcd_gap = 8)
}
\arguments{
\item{x}{a numeric vector.}

\item{neighbour}{find local extreme within \eqn{+/- neighbour}.}

\item{mincycle}{numeric, one cycle must have at least \code{mincycle} months, if a monthly time series.}

\item{phase}{numeric, a phase of one cycle must have length of \code{phase} months, phase means peak to peak or trough to trough.}

\item{extr_limit}{numeric, some values will be regarded as extreme who out range of \eqn{+/- extr_limit} standard deviation of ratio of irregular and original series.}

\item{end_num}{peak or trough should be apart form either end of series by at least \code{end_num} months.}

\item{mcd_gap}{max gap for calculate ratio when getting \emph{MCD} value, default is 8.}
}
\value{
\itemize{
 \item \code{isturn} a vector stores turns information of series. 1 means it is a peak turn, -1 means it is trough turn, 0 means neither peak nor trough.
 \item \code{isturn_old} a vector stores turns information for last setup.
 \item \code{peaktrough_idx} stores index of peak or trough turns of original series.
 \item \code{peak_idx} peak index.
 \item \code{trough_idx} trough index.
}
}
\description{
\code{bbturns} returns turns information of a time series, selected by B-B algorithm.
}
\details{
B-B algorithm proposed by Gerhard Bry and Charlotte Boschan in 1971, see \url{http://papers.nber.org/books/bry_71-1}.
}
\examples{
data(importexport)
bbturns(importexport$ImEx)
}
\author{
Hejun Liu, Beijing \email{liuhejunlj@163.com}
}
