% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmAverage.R
\name{cmAverage}
\alias{cmAverage}
\title{centralized moving average}
\usage{
cmAverage(x, frequency = numeric())
}
\arguments{
\item{x}{a number series.}

\item{frequency}{get a \code{frequency} period moving average.}
}
\examples{
x = c(2,3,4,5,2,7)
cmAverage(x, 3)
x = c(2,3,4,5,2,7)
cmAverage(x,4)
}
