% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrmadj.R
\name{extrmadj}
\alias{extrmadj}
\title{adjuste extreme values}
\usage{
extrmadj(x, limit = 3.5, subs_method = c("spencer", "nearest"))
}
\arguments{
\item{x}{numeric vector.}

\item{limit}{threshold for recognizing extreme value.}

\item{subs_method}{indicate method for substituting extreme values, \code{spencer} or \code{nearest}, 
the former means substituted by \emph{spencer} value, latter means substituted by mean of nearest 
none extreme values.}
}
\description{
Adjusting extreme values of series.
}
\examples{

##
usad1 = USAccDeaths
usad1[2:5] = mean(USAccDeaths) * 3  # first obs will become a extreme since the trend has been changed by 2:5
(ext1 = extrmadj(usad1))
plot(ext1$extr_infr$x, type = "l") # original
lines(ext1$x.nextr, type = "l", col = "red") # substituted
## 
usad2 = USAccDeaths
usad2[15] = - mean(USAccDeaths) * 3
(ext2 = extrmadj(usad2))
plot(ext2$extr_infr$x, type = "l") # original
lines(ext2$x.nextr, type = "l", col = "red") # substituted

## substituted by mean of nearest non-extrmadj
(ext3 = extrmadj(usad2, subs_method = "near"))
plot(ext3$extr_infr$x, type = "l") # original
lines(ext3$x.nextr, type = "l", col = "red") # substituted

}
