% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importexport.R
\docType{data}
\name{importexport}
\alias{importexport}
\title{Monthly year on year growth of imports and exports data of China.}
\format{A 3 x 254 data.frame}
\usage{
importexport
}
\description{
This data set is a format of data.frame, contains monthly  year on year growth of imports-exports, 
imports, exports from 1995.8 to 2016.9, there are altogether 254 observations.
}
\keyword{datasets}
