% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpMAverage.R
\name{SpMAverage}
\alias{SpMAverage}
\title{Spencer moving average}
\usage{
SpMAverage(x)
}
\arguments{
\item{x}{numeric vector.}
}
\description{
return a 15 period moving average.
}
\examples{
x = c(1:20)
SpMAverage(x)
}
